+++ addons/website/views/website_navbar_templates.xml
@@ -60,13 +60,6 @@
                             <li><a data-action="promote-current-page" href="#" title="Promote page on the web">Optimize SEO</a></li>
                         </ul>
                     </li>
-                    <!-- Help -->
-                    <li class="dropdown" groups="base.group_no_one">
-                        <a class="dropdown-toggle waves" data-toggle="dropdown" href="#" id="help-menu-button">Help</a>
-                        <ul class="dropdown-menu" id="help-menu" role="menu">
-                            <li class="dropdown-header">HELP &amp; TUTORIALS</li>
-                        </ul>
-                    </li>
                 </ul>
 
                 <ul class="o_menu_systray hidden-xs">
